+++ src/protocol/mod.rs
@@ -3,14 +3,12 @@ mod batch;
 mod check_zero;
 pub mod context;
 pub mod malicious;
-mod maliciously_secure_mul;
 mod modulus_conversion;
 pub mod prss;
 mod reveal;
 mod reveal_additive_binary;
-mod securemul;
 pub mod sort;
-pub(crate) mod mul;
+pub mod mul;
 
 #[cfg(debug_assertions)]
 use std::{
